requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-speech==2.26.0
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.0
googletrans==4.0.0rc1
python-dotenv==1.0.1
scikit-learn==1.4.2
